+++ src/core/elyxion.h
@@ -9,9 +9,9 @@
 
 // Version info
 #define ELYXION_VERSION_MAJOR 1
-#define ELYXION_VERSION_MINOR 0
+#define ELYXION_VERSION_MINOR 1
 #define ELYXION_VERSION_PATCH 0
-#define ELYXION_VERSION_STRING "1.0.0"
+#define ELYXION_VERSION_STRING "1.1.0"
 
 namespace elyxion {